Refactor test and prepare for injecting constants
